refactor SignInPage
